Various fixes
Bugfixes:

Fixed syntax errors;
Fixed several typographical errors. c was renamed to s in some places but not in others;

Made the bot less trigger-happy. For comments, it will only trigger if the commenter is actually asking for something and if the body is less than 50 characters long, unless explicitly summoned. For posts, it only checks the title and flair.