Bugfixes regarding TrialBot
The TrialBot implementation (discordtest.py) raises IndexError when trying to query a name that doesn't exist. This didn't use to happen. Either the TrialBot specification has changed without my knowledge or something wasn't properly merged when dealing with Pyro's merge request. Or the third and most likely case: because of my own incompetence. Also changed a fews reference to config.py to correctly point to settings.py and secrets.py